Re-organize src files
